SAP_clone/backend/requirements.txt
# FastAPI and web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1

# Validation and serialization
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0

# HTTP client
httpx>=0.26.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
hypothesis>=6.92.2

# Observability
prometheus-client>=0.19.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
